Rearrange map components into seperate files
