+++ media/src/project/project-map.tsx
@@ -5,6 +5,9 @@ import 'mapbox-gl/dist/mapbox-gl.css';
 // Deck.gl
 import DeckGL from '@deck.gl/react';
 
+import { ProjectMapSidebar, ProjectMapSidebarProps } from './project-map-sidebar';
+import { LayerProps } from './layer';
+
 const authedFetch = (url: string, method: string, data: any) => {
     let csrf = (document.getElementById('csrf-token') as any).getAttribute('content');
     return fetch(url,{
@@ -182,97 +185,4 @@ export const ProjectMap = () => {
             </>
         );
     }
-};
-
-interface ProjectMapSidebarProps {
-    title: string;
-    description: string;
-    layers: LayerProps[]
-    addLayer(layerTitle: string): any;
-    deleteLayer(pk: number): any;
-    updateLayer(pk: number, title: string): any;
-}
-
-export const ProjectMapSidebar = (
-    {title, description, layers, addLayer, deleteLayer, updateLayer}: ProjectMapSidebarProps) => {
-
-    const [newLayerTitle, setNewLayerTitle] = useState<string>('');
-
-    const handleNewLayerTitle = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setNewLayerTitle(e.target.value);
-    };
-
-    const handleCreateLayer = (e: React.FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-        addLayer((e.currentTarget.elements[0] as HTMLInputElement).value);
-        setNewLayerTitle('');
-    };
-
-    return (
-        <div id='project-map-sidebar'>
-            <h2>{title}</h2>
-            <p>{description}</p>
-            <h3>Layers</h3>
-            <form onSubmit={handleCreateLayer}
-                className="needs-validation" noValidate >
-                <div className="form-group">
-                    <label>Layer Title:
-                        <input id="new-layer-title"
-                            value={newLayerTitle}
-                            onChange={handleNewLayerTitle}
-                            className="form-control" type="text"/>
-                    </label>
-                </div>
-                <input type='submit'
-                    className='btn btn-primary' value={'Add Layer'}/>
-            </form>
-            {layers && layers.map(
-                (layer, idx) => {return (<Layer {...layer} deleteLayer={deleteLayer} updateLayer={updateLayer} key={idx} />);})}
-        </div>
-    );
-};
-
-interface LayerProps {
-    title: string;
-    pk: number;
-    content_object: string; // The API URL to the parent project/response
-    deleteLayer(pk: number): any;
-    updateLayer(pk: number, title: string): any;
-}
-
-export const Layer = (layerData: LayerProps)=> {
-    const [updatedLayerTitle, setUpdatedLayerTitle] = useState<string>(layerData.title);
-
-    const handleUpdatedLayerTitle = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setUpdatedLayerTitle(e.target.value);
-    };
-    const handleUpdateLayer = (e: React.FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-        layerData.updateLayer(layerData.pk, updatedLayerTitle);
-    };
-
-    const handleDeleteLayer = (e: React.FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-        layerData.deleteLayer(layerData.pk);
-    };
-    return (
-        <div className="project-map-sidebar-layer">
-            <span className="font-weight-bold">{layerData.title}</span>
-            <p>{layerData.content_object}</p>
-            <form onSubmit={handleUpdateLayer}>
-                <label>Layer Title:
-                    <input id={`update-layer-title-${layerData.pk}`}
-                        value={updatedLayerTitle}
-                        onChange={handleUpdatedLayerTitle}
-                        className="form-control" type="text"/>
-                </label>
-                <input type='submit'
-                    className='btn btn-primary' value={'Edit Layer'}/>
-            </form>
-            <form onSubmit={handleDeleteLayer}>
-                <input type='submit'
-                    className='btn btn-danger' value={'Delete Layer'}/>
-            </form>
-        </div>
-    );
 };
